added email notify support

+++ task_lister/models.py
@@ -10,7 +10,7 @@ class User(UserMixin , db.Model):
     id = db.Column(db.Integer, primary_key = True)
     username = db.Column(db.String(64), index = True, unique = True)
     email = db.Column(db.String(120), index = True, unique = True)
-    mob = db.Column(db.Integer,unique = True)
+    mob = db.Column(db.BigInteger(),unique = True)
     password_hash = db.Column(db.String(120))
     tasks = db.relationship('Task', backref = 'user', lazy = 'dynamic', cascade = 'all, delete-orphan')
